Preliminary efforts -- MVP for Python RapidAnnotator
This is a very minimal version of the Python port of Rapid
Annotator, featuring no user account support or administrative
tools, as well as only supporting locally-hosted images.  It
does, however, feature the basic underlying behavior of Rapid
Annotator (annotation of images with a user configurable set
of responses), and store the data in a database rather than
in static text files.